+++ src/components/util/search-filters.js
@@ -123,97 +123,48 @@ export function getStatusAndUnselectHrefIfSelectedOrOmittedFromResponseFilters(t
     //workaround: '!=' condition adds '!' to the end of facet.field in StaticSingleTerm, we remove it
     const field = facet.field.endsWith('!') ? facet.field.slice(0, -1) : facet.field;
 
-    const isRange = facet.aggregation_type && ['range', 'date_histogram', 'histogram'].indexOf(facet.aggregation_type) > -1;
-    let i, filter, parts, retHref = '';
+    let i, filter, parts, retHref = '', found = false, status = "selected";
 
-    // THE CONTENTS UNDER THIS IF CONDITION WILL CHANGE ONCE WE CREATE NEW 'RANGE' FACET COMPONENT
-    if (facet.aggregation_type && ['range', 'date_histogram', 'histogram'].indexOf(facet.aggregation_type) > -1) {
-        var toFilter, fromFilter;
-
-        if (facet.aggregation_type === 'range'){
-            toFilter    = _.findWhere(filters, { 'field' : field + '.to',   'term' : term.to }),
-            fromFilter  = _.findWhere(filters, { 'field' : field + '.from', 'term' : term.from });
-        } else if (facet.aggregation_type === 'date_histogram'){
-            var interval = getDateHistogramIntervalFromFacet(facet) || 'month',
-                toDate = moment.utc(term.key);
-            toDate.add(1, interval + 's');
-            toFilter    = _.findWhere(filters, { 'field' : field + '.to',   'term' : toDate.format().slice(0,10) }),
-            fromFilter  = _.findWhere(filters, { 'field' : field + '.from', 'term' : term.key });
-        } else {
-            throw new Error('Histogram not currently supported.');
-            // Todo: var interval = ....
+    if (facet.aggregation_type === "stats"){
+        for (i = 0; i < filters.length; i++) {
+            filter = filters[i];
+            const filterTermValue = parseFloat(filter.term);
+            if (isNaN(filterTermValue)) continue;
+            if ((filter.field === field) && (filterTermValue === term.key)){
+                found = true;
+                break;
+            }
         }
-
-        if (toFilter && !fromFilter){
-            parts = url.parse(toFilter['remove']);
+        if (found) {
+            parts = url.parse(filter.remove);
             if (includePathName) {
                 retHref += parts.pathname;
             }
             retHref += parts.search;
-            return { 'status': 'selected', 'href': retHref };
-        } else if (!toFilter && fromFilter){
-            parts = url.parse(fromFilter['remove']);
-            if (includePathName) {
-                retHref += parts.pathname;
-            }
-            retHref += parts.search;
-            return { 'status': 'selected', 'href': retHref };
-        } else if (toFilter && fromFilter){
-            var partsFrom   = url.parse(fromFilter['remove'], true),
-                partsTo     = url.parse(toFilter['remove'], true),
-                partsFromQ  = partsFrom.query,
-                partsToQ    = partsTo.query,
-                commonQs    = {};
-
-            _.forEach(_.keys(partsFromQ), function(qk){
-                if (typeof partsToQ[qk] !== 'undefined'){
-                    if (Array.isArray(partsToQ[qk]) || Array.isArray(partsFromQ[qk])){
-                        var a1, a2;
-                        if (Array.isArray(partsToQ[qk])) {
-                            a1 = partsToQ[qk];
-                        } else {
-                            a1 = [partsToQ[qk]];
-                        }
-                        if (Array.isArray(partsFromQ[qk])) {
-                            a2 = partsFromQ[qk];
-                        } else {
-                            a2 = [partsFromQ[qk]];
-                        }
-                        commonQs[qk] = _.intersection(a1, a2);
-                    } else {
-                        commonQs[qk] = partsToQ[qk];
-                    }
-                }
-            });
-
-            retHref = '?' + queryString.stringify(commonQs);
-            if (includePathName) {
-                retHref += partsFrom.pathname;
-            }
-            return { 'status': 'selected', 'href': retHref };
+            return { status, 'href': retHref };
         }
-
     } else {
         // Terms
         for (i = 0; i < filters.length; i++) {
             filter = filters[i];
-            if (filter.field == field && filter.term == term.key) {
-                parts = url.parse(filter.remove);
-                if (includePathName) {
-                    retHref += parts.pathname;
-                }
-                retHref += parts.search;
-                return { 'status': 'selected', 'href': retHref };
-            }
-            else if (filter.field.endsWith('!') && filter.field.slice(0, -1) == field && filter.term == term.key) {
-                parts = url.parse(filter.remove);
-                if (includePathName) {
-                    retHref += parts.pathname;
-                }
-                retHref += parts.search;
-                return { 'status': 'omitted', 'href': retHref };
+            if (filter.field === field && filter.term === term.key) {
+                found = true;
+                break;
+            } else if (filter.field.endsWith('!') && filter.field.slice(0, -1) === field && filter.term === term.key) {
+                found = true;
+                status = "omitted";
+                break;
             }
         }
+        if (found) {
+            parts = url.parse(filter.remove);
+            if (includePathName) {
+                retHref += parts.pathname;
+            }
+            retHref += parts.search;
+            return { status, 'href': retHref };
+        }
+
     }
     return { 'status': 'none', 'href': null };
 }
